Seperated types to prevent circular dependencies
