+++ apps/services/search-indexer/src/migrate/elastic.ts
@@ -2,11 +2,8 @@ import fs from 'fs'
 import { logger } from '@island.is/logging'
 import { ElasticService } from '@island.is/api/content-search'
 import { AwsEsPackage } from './aws'
-import {
-  IndexingModule,
-  IndexingService,
-  SyncOptions,
-} from '@island.is/elastic-indexing'
+import { SyncOptions } from '@island.is/elastic-indexing/types'
+import { IndexingModule, IndexingService } from '@island.is/elastic-indexing'
 import { NestFactory } from '@nestjs/core'
 
 const esService = new ElasticService()
