+++ libs/api/domains/cms/src/lib/search/cmsSync.service.ts
@@ -5,9 +5,10 @@ import { hashElement } from 'folder-hash'
 import {
   ContentSearchImporter,
   MappedData,
+  SearchIndexes,
   SyncOptions,
   SyncResponse,
-} from '@island.is/elastic-indexing'
+} from '@island.is/elastic-indexing/types'
 import { ArticleSyncService } from './importers/article.service'
 import { ContentfulService } from './contentful.service'
 import { LifeEventsPageSyncService } from './importers/lifeEventsPage.service'
@@ -15,7 +16,7 @@ import { ArticleCategorySyncService } from './importers/articleCategory.service'
 import { NewsSyncService } from './importers/news.service'
 import { AboutPageSyncService } from './importers/aboutPage.service'
 import { Entry } from 'contentful'
-import { ElasticService, SearchIndexes } from '@island.is/api/content-search'
+import { ElasticService } from '@island.is/api/content-search'
 
 export interface PostSyncOptions {
   folderHash: string
