+++ libs/elastic-indexing/src/index.ts
@@ -1,3 +1,2 @@
 export * from './lib/indexing.module'
 export * from './lib/indexing.service'
-export * from './lib/types'
