+++ libs/elastic-indexing/src/lib/dto/syncInput.input.ts
@@ -1,6 +1,6 @@
+import { SearchIndexes } from '@island.is/elastic-indexing/types'
 import { Field, InputType } from '@nestjs/graphql'
 import { IsOptional, IsString } from 'class-validator'
-import { SearchIndexes } from '@island.is/api/content-search'
 
 @InputType()
 export class SyncInput {
