+++ libs/elastic-indexing/types/.eslintrc
@@ -1,0 +1,1 @@
+{ "extends": "../../../.eslintrc", "rules": {}, "ignorePatterns": ["!**/*"] }
