+++ libs/elastic-indexing/types/jest.config.js
@@ -1,0 +1,14 @@
+module.exports = {
+  name: 'shared-types',
+  preset: '../../../jest.config.js',
+  globals: {
+    'ts-jest': {
+      tsConfig: '<rootDir>/tsconfig.spec.json',
+    },
+  },
+  transform: {
+    '^.+\\.[tj]sx?$': 'ts-jest',
+  },
+  moduleFileExtensions: ['ts', 'tsx', 'js', 'jsx'],
+  coverageDirectory: '../../../coverage/libs/elastic-indexing/types',
+}
