+++ workspace.json
@@ -1758,6 +1758,35 @@
         }
       }
     },
+    "elastic-indexing-types": {
+      "root": "libs/elastic-indexing/types",
+      "sourceRoot": "libs/elastic-indexing/types/src",
+      "projectType": "library",
+      "schematics": {},
+      "architect": {
+        "lint": {
+          "builder": "@nrwl/linter:lint",
+          "options": {
+            "linter": "eslint",
+            "tsConfig": [
+              "libs/elastic-indexing/types/tsconfig.lib.json",
+              "libs/elastic-indexing/types/tsconfig.spec.json"
+            ],
+            "exclude": [
+              "**/node_modules/**",
+              "!libs/elastic-indexing/types/**/*"
+            ]
+          }
+        },
+        "test": {
+          "builder": "@nrwl/jest:jest",
+          "options": {
+            "jestConfig": "libs/elastic-indexing/types/jest.config.js",
+            "passWithNoTests": true
+          }
+        }
+      }
+    },
     "email-service": {
       "root": "libs/email-service",
       "sourceRoot": "libs/email-service/src",
